Add more assertions, rename test folder
